synaptics: Add first skeleton of a test
Unfortunately it fails when run along with a real Xserver as
the device gets grabbed by the running Xserver and thus prevents
the test Xserver from accessing the evemu device.

This test needs to be run on a machine without an existing running
Xserver. Committing to git even though this is far from being
complete yet for further testing on a remote system.

Signed-off-by: Olivier Fourdan <ofourdan@redhat.com>