Drop the parameter interface from the default InputDriverTest class
Having this interface defined in the default class leads to ambiguity when
we want to define a parameter interface of a different type in a subclass.
So instead, drop it from the main parent class, provide a simple SetUp()
that calls the one subclasses can use if they want to pass something into
SetUpConfigAndLog().

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>