wacom: use the InputDriverTest base class
We're not really doing much different here in terms of setup, so we can
re-use most of the code from the base class.

A real change is now that the device is initialised before the server
starts, so we need to check if the device is already present before we worry
about hierarchy events.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>