server: add a test for owner event field in XIGrabButton()
The logic is as follow:

If owner_events is True, the event is reported if selected by
the window's event mask. if that mask isn't set, the event is
reported if the grab mask is set.
If owner_events is False, the event is only reported if
the grab mask is set and the window event mask is ignored.