common: use -noreset by default
xorg-gtest essentially prevents avoiding server reset. WaitForConnection()
is built to always reset and even if it didn't, Test::SetUp() opens its own
connection anyway, not allowing for only one display connection that can be
used to force a server reset.

Until that mess is updated, use -noreset here since none of the (current!)
tests want a reset, it gains us some speedup and noreset is easy enough to
grep for if someone needs to know why the server doesn't reset.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>