server/misc: add a bunch of idletime sync counter test
Definitely broken upstream. If a NegativeTransition counter is set on the
idletimer, an AlarmNotify is sent when the idletimer goes below that. Once
the alarm is sent, the server calls ComputeBracket. That should, but doesn't
set the negative transition again because idletime < threshold for a
negative transition (i.e. it's skipped).

The idletimer can then go past the threshold, and back to 0 because nothing
is ever set.

Since the brackets are re-computed after each alarm, if there is one more
alarm that fires the negative transition may be re-institutated after that
alarm.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>