common: add an exception to timeout handling
Depending when the timeout happens, the FAIL won't actually stop the test.
e.g. if we timeout during a select()/poll()/etc. we just return to the syscall
after the signal handler is done and hang there.

An exception would stop the test, but only works once, because throwing an
exception in a signal handler is not great. But then again, it beats having a
test stall indefinitely. No good solutions here because we only have one
process and googletest doesn't have any timeout handling.

So, let's choose the most cringeworthy path: try to FAIL, if that (haha)
fails, then throw an exception. This at least hobbles along on to the next test
instead of hanging at the first failure. With any luck we only have one
real test failure that needs an exception, if not we're SOOL anyway. It's much
like taking a gun with two bullets, shooting one foot and if that doesn't
work, blow away the other one too to see if we can walk away from the carnage.
Fun to be had by everyone.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>